Refactor Windows sandbox unified exec sessions
